+++ utils/common/polylib.c
@@ -17,7 +17,7 @@ winding_t	*AllocWinding (int points)
 
 	s = sizeof(vec_t)*3*points + sizeof(int);
 	w = malloc (s);
-	memset (w, 0, s); 
+	memset (w, 0, s);
 	return w;
 }
 
@@ -128,7 +128,7 @@ winding_t *BaseWindingForPlane (vec3_t normal, float dist)
 	vec_t	max, v;
 	vec3_t	org, vright, vup;
 	winding_t	*w;
-	
+
 // find the major axis
 
 	max = -BOGUS_RANGE;
@@ -144,48 +144,48 @@ winding_t *BaseWindingForPlane (vec3_t normal, float dist)
 	}
 	if (x==-1)
 		Error ("BaseWindingForPlane: no axis found");
-		
-	VectorCopy (vec3_origin, vup);	
+
+	VectorCopy (vec3_origin, vup);
 	switch (x)
 	{
 	case 0:
 	case 1:
 		vup[2] = 1;
-		break;		
+		break;
 	case 2:
 		vup[0] = 1;
-		break;		
+		break;
 	}
 
 	v = DotProduct (vup, normal);
 	VectorMA (vup, -v, normal, vup);
 	VectorNormalize (vup);
-		
+
 	VectorScale (normal, dist, org);
-	
+
 	CrossProduct (vup, normal, vright);
-	
+
 	VectorScale (vup, 8192, vup);
 	VectorScale (vright, 8192, vright);
 
 // project a really big	axis aligned box onto the plane
 	w = AllocWinding (4);
-	
+
 	VectorSubtract (org, vright, w->p[0]);
 	VectorAdd (w->p[0], vup, w->p[0]);
-	
+
 	VectorAdd (org, vright, w->p[1]);
 	VectorAdd (w->p[1], vup, w->p[1]);
-	
+
 	VectorAdd (org, vright, w->p[2]);
 	VectorSubtract (w->p[2], vup, w->p[2]);
-	
+
 	VectorSubtract (org, vright, w->p[3]);
 	VectorSubtract (w->p[3], vup, w->p[3]);
-	
+
 	w->numpoints = 4;
-	
-	return w;	
+
+	return w;
 }
 
 /*
@@ -197,8 +197,8 @@ winding_t	*CopyWinding (winding_t *w)
 {
 	int			size;
 	winding_t	*c;
-	
-	size = (int)((winding_t *)0)->p[w->numpoints];
+
+	size = (intptr_t)((winding_t *)0)->p[w->numpoints];
 	c = malloc (size);
 	memcpy (c, w, size);
 	return c;
@@ -222,7 +222,7 @@ void	ClipWinding (winding_t *in, vec3_t normal, vec_t dist,
 	vec3_t	mid;
 	winding_t	*f, *b;
 	int		maxpts;
-	
+
 	counts[0] = counts[1] = counts[2] = 0;
 
 // determine sides for each point
@@ -243,7 +243,7 @@ void	ClipWinding (winding_t *in, vec3_t normal, vec_t dist,
 	}
 	sides[i] = sides[0];
 	dists[i] = dists[0];
-	
+
 	*front = *back = NULL;
 
 	if (!counts[0])
@@ -262,11 +262,11 @@ void	ClipWinding (winding_t *in, vec3_t normal, vec_t dist,
 
 	*front = f = AllocWinding (maxpts);
 	*back = b = AllocWinding (maxpts);
-		
+
 	for (i=0 ; i<in->numpoints ; i++)
 	{
 		p1 = in->p[i];
-		
+
 		if (sides[i] == SIDE_ON)
 		{
 			VectorCopy (p1, f->p[f->numpoints]);
@@ -275,7 +275,7 @@ void	ClipWinding (winding_t *in, vec3_t normal, vec_t dist,
 			b->numpoints++;
 			continue;
 		}
-	
+
 		if (sides[i] == SIDE_FRONT)
 		{
 			VectorCopy (p1, f->p[f->numpoints]);
@@ -289,10 +289,10 @@ void	ClipWinding (winding_t *in, vec3_t normal, vec_t dist,
 
 		if (sides[i+1] == SIDE_ON || sides[i+1] == sides[i])
 			continue;
-			
+
 	// generate a split point
 		p2 = in->p[(i+1)%in->numpoints];
-		
+
 		dot = dists[i] / (dists[i]-dists[i+1]);
 		for (j=0 ; j<3 ; j++)
 		{	// avoid round off error when possible
@@ -303,13 +303,13 @@ void	ClipWinding (winding_t *in, vec3_t normal, vec_t dist,
 			else
 				mid[j] = p1[j] + dot*(p2[j]-p1[j]);
 		}
-			
+
 		VectorCopy (mid, f->p[f->numpoints]);
 		f->numpoints++;
 		VectorCopy (mid, b->p[b->numpoints]);
 		b->numpoints++;
 	}
-	
+
 	if (f->numpoints > maxpts || b->numpoints > maxpts)
 		Error ("ClipWinding: points exceeded estimate");
 	if (f->numpoints > MAX_POINTS_ON_WINDING || b->numpoints > MAX_POINTS_ON_WINDING)
@@ -352,13 +352,13 @@ void CheckWinding (winding_t *w)
 
 	if (w->numpoints < 3)
 		Error ("CheckFace: %i points",w->numpoints);
-	
+
 	area = WindingArea(w);
 	if (area < 1)
 		Error ("CheckFace: %f area", area);
 
 	WindingPlane (w, facenormal, &facedist);
-	
+
 	for (i=0 ; i<w->numpoints ; i++)
 	{
 		p1 = w->p[i];
@@ -368,24 +368,24 @@ void CheckWinding (winding_t *w)
 				Error ("CheckFace: BUGUS_RANGE: %f",p1[j]);
 
 		j = i+1 == w->numpoints ? 0 : i+1;
-		
+
 	// check the point is on the face plane
 		d = DotProduct (p1, facenormal) - facedist;
 		if (d < -ON_EPSILON || d > ON_EPSILON)
 			Error ("CheckFace: point off plane");
-	
+
 	// check the edge isn't degenerate
 		p2 = w->p[j];
 		VectorSubtract (p2, p1, dir);
-		
+
 		if (VectorLength (dir) < ON_EPSILON)
 			Error ("CheckFace: degenerate edge");
-			
+
 		CrossProduct (facenormal, dir, edgenormal);
 		VectorNormalize (edgenormal);
 		edgedist = DotProduct (p1, edgenormal);
 		edgedist += ON_EPSILON;
-		
+
 	// all other points must be on front side
 		for (j=0 ; j<w->numpoints ; j++)
 		{
